requirements.txt
appnope==0.1.0
argh==0.26.2
backcall==0.1.0
decorator==4.3.0
flake8==3.5.0
greenlet==0.4.13
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
mccabe==0.6.1
msgpack==0.5.6
mypy==0.610
neovim==0.2.6
parso==0.2.1
pathtools==0.1.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyserial==3.4
PyYAML==3.12
rope==0.10.7
simplegeneric==0.8.1
six==1.11.0
smbus2==0.2.1
traitlets==4.3.2
typed-ast==1.1.0
watchdog==0.8.3
wcwidth==0.1.7
yapf==0.22.0
